Cambia nombres a estándar closes #27 refs #26
