+++ main.js
@@ -57,7 +57,7 @@ app.on('activate', function() {
 
 
 var ipc = require('electron').ipcMain;
-const Downloader = require("./downloader-class.js");
+const Downloader = require("./downloader.js");
 var updater = require("./mainWindowUpdater.js");
 
 var downloaderInstance = new Downloader();
